+++ game-rest/src/main/java/com/kevinguanchedarias/owgejava/response/UnitTypeResponse.java
@@ -1,0 +1,14 @@
+package com.kevinguanchedarias.owgejava.response;
+
+import com.kevinguanchedarias.owgejava.dto.UnitTypeDto;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@EqualsAndHashCode(callSuper = false)
+public class UnitTypeResponse extends UnitTypeDto {
+    boolean used;
+}